refactor: add new structure to directories
