+++ backend/src/apps/books/books.service.ts
