+++ backend/src/apps/books/dtos/index.ts
