+++ backend/src/apps/books/schemas/books.schema.ts
